+++ js/full-photo-click-handler.js
